requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.9.2
pydantic-settings==2.1.0
email-validator==2.1.0
supabase==2.8.1
httpx==0.27.2
python-multipart==0.0.6
openai==1.3.7
beautifulsoup4==4.12.2
lxml>=5.0.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0
apscheduler==3.10.4
pytz==2023.3
firecrawl-py==0.0.16
tenacity>=8.2.0